.control/controlman/pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "versioningit"]
build-backend = "setuptools.build_meta"


# ----------------------------------------- setuptools -------------------------------------------
[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
where = ["src"]
namespaces = true


# ----------------------------------------- Project Metadata -------------------------------------
#
[project]
version = "0.1.0"
name = "ControlMan"
dependencies = [
    "packaging >=23.2,<24",
    "trove-classifiers >=2024.9.12",
    "jsonschema >=4.23,<5",
    "referencing >=0.35.1",
    "jsonpath-ng >=1.6.1,<2",
    "ruamel.yaml >=0.18.6,<0.19",
    "Jinja2 >=3,<4",
    "PyLinks >=0.1,<0.2",
    "LoggerMan >=0.1,<0.2",
    "PySerials >=0.1,<0.2",
    "GitTidy >=0.1,<0.2",
    "PkgData >=0.1,<0.2",
    "PyShellMan >=0.1,<0.2",
    "PySyntax >=0.1,<0.2",
    "ExceptionMan >=0.1,<0.2",
    "MDit >=0.1,<0.2",
    "JSONSchemata >=0.1,<0.2",
    "VersionMan >=0.1,<0.2",
    "HTMP >=0.1,<0.2",
    "LicenseMan >=0.1,<0.2",
]
requires-python = ">=3.10"
